refactor: refine sip rtp runtime ownership
