+++ api/assistant-api/sip/runtime/types_test.go
@@ -61,6 +61,54 @@ func TestSIPRequestContextResolveRouteNilContext(t *testing.T) {
 	assert.Nil(t, route)
 }
 
+func TestRTPConfigValidateRejectsNil(t *testing.T) {
+	var config *RTPConfig
+
+	assert.ErrorIs(t, config.Validate(), errRTPConfigRequired)
+}
+
+func TestRTPConfigValidateReturnsSentinelErrors(t *testing.T) {
+	tests := []struct {
+		name    string
+		config  *RTPConfig
+		wantErr error
+	}{
+		{
+			name:    "missing local ip",
+			config:  &RTPConfig{},
+			wantErr: errRTPLocalIPRequired,
+		},
+		{
+			name: "blank local ip",
+			config: &RTPConfig{
+				LocalIP: " ",
+			},
+			wantErr: errRTPLocalIPRequired,
+		},
+		{
+			name: "invalid local port",
+			config: &RTPConfig{
+				LocalIP:   "127.0.0.1",
+				LocalPort: rtpMaxPort + 1,
+			},
+			wantErr: errRTPInvalidLocalPort,
+		},
+		{
+			name: "missing port range",
+			config: &RTPConfig{
+				LocalIP: "127.0.0.1",
+			},
+			wantErr: errRTPPortRangeRequired,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			assert.ErrorIs(t, test.config.Validate(), test.wantErr)
+		})
+	}
+}
+
 func TestParseConfigFromVault_NilCredential(t *testing.T) {
 	_, err := ParseConfigFromVault(nil)
 	require.Error(t, err)
